refactor: extract lookup
